Fix mistake where total-mapped-count is not calculated correctly and more accounts than wanted are mapped.
Add Rarity and ItemSet to store the rarity of the item and the ItemSet / Collection it belongs to (such as Gods & Monsters Collection).

Map some values in the HttpInventoryResponse more intelligently.

Distribute IDs to be mapped more evenly for performance. Log more exceptions to be able to debug failed requests.

Add two missing values to Exterior enum.