Overhaul of the way items are stored. Completely remove Transient Objects since they aren't needed with Hibernate.
Overhaul the way Items are typed -> Item Types now include attributes such as the item being StatTrak or Souvenir, the Rarity, Exterior, etc. This massively increases the performance of many operations. The database file size is also decreased in size by almost half.

Also fix a mistake where items with the same Class-ID but different instance-ids are sometimes classified incorrectly.

Rewrite all database queries to work with the new entities. Update all writers to work with the new standard.

Rework the way stickers with a "," in their name are repaired for gun-placements. Use fully automatic and expandable version for the future.

Remove MANY null-pointers for very small databases, where some item-sets or items are simply not found yet.

Remove the statTrak and souvenir booleans and replace them with a SpecialItemType enum.